Map enhancements: split GenericMapPythonVisitor into its own file map.hpp, add visitor arg to ::expose() with overloads
